Added Assets Page

+++ src/client/Client.Infrastucture/Models/AssetToken.cs
@@ -7,7 +7,7 @@ namespace Client.Infrastructure.Models
     {
         public UInt160 AssetHash { get; set; }
         public string Symbol { get; set; }
-        public BigInteger Balance { get; set; }
+        public decimal Balance { get; set; }
         public byte Decimals { get; set; }
     }
 }
